refactor(extension): move coverage utils to break circular import
